requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
aiosqlite==0.20.0
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
jinja2==3.1.4
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2
PyJWT==2.9.0
bcrypt==4.2.0
